+++ Sandbox/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(SOURCES
 	src/SandboxApp.cpp
+    src/ExampleLayer.cpp
 	)
 
 add_executable(sandbox
